Refactor ast.rs
- move embedding code to embed.rs
- reorder methods
- shorten Debug and Display methods
